DM: lib: ugly ugly ugly ASL boost 1.53 initializer workaround

+++ modules/ftm/include/minicommand.h
@@ -50,7 +50,7 @@ public:
   void serialise(uint8_t* b) const {
     MiniCommand::serialise(b);
     writeLeNumberToBeBytes(b + (ptrdiff_t)T_CMD_FLOW_DEST, this->destAdr);
-    writeLeNumberToBeBytes(b + (ptrdiff_t)T_CMD_RES, 0);   
+    writeLeNumberToBeBytes(b + (ptrdiff_t)T_CMD_RES, (uint32_t)0);   
   };
 
 };
